Updates for Svc/PrmDb corresponding to UT cmake update for protected/private by updating driver name and updating PrmDbImplTester to PrmDbTester
